+++ src/models/__init__.py
@@ -1,3 +1,4 @@
-from .app_project_model import AppProjectModel
+from .app_project_model import AppProjectModel, AppProjectDTO
+from .base import Base
 
-__all__ = [AppProjectModel]
+__all__ = [AppProjectModel, AppProjectDTO, Base]