+++ CHANGELOG.md
@@ -20,3 +20,4 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Docker Compose configuration for multi-container orchestration.
 - Initial deployment scripts and example files for hosting the CV site on a production server.
 - AppSettings [8-add-get-projects]
+- AppProjectDTO [8-add-get-projects]